refactor: separate text alignment controls into a new component
