+++ src/shared/markdown/MdxEditorTableControls.tsx
@@ -13,18 +13,15 @@ import {
 } from '@chakra-ui/react'
 import {
   ButtonWithTooltip,
-  editorInFocus$,
   editorInTable$,
   insertTable$,
   readOnly$,
-  rootEditor$,
   useCellValue,
   usePublisher,
 } from '@mdxeditor/editor'
 import { t } from 'i18next'
-import { type LexicalEditor, FORMAT_ELEMENT_COMMAND } from 'lexical'
 import { type KeyboardEvent, useState } from 'react'
-import { PiTable, PiTextAlignCenter, PiTextAlignLeft, PiTextAlignRight } from 'react-icons/pi'
+import { PiTable } from 'react-icons/pi'
 
 import { Body } from '@/shared/components/typography'
 import { useDebounce } from '@/shared/hooks/useDebounce.ts'
@@ -39,14 +36,6 @@ const tableBoxes = Array.from({ length: TABLE_ROW_COUNT }, (_, rowIndex) =>
 type tablePosition = {
   row: number
   column: number
-}
-
-const getFocusedEditor = (editorInFocusRef: unknown, rootEditor: LexicalEditor | null): LexicalEditor | null => {
-  if (editorInFocusRef && typeof (editorInFocusRef as LexicalEditor).dispatchCommand === 'function') {
-    return editorInFocusRef as LexicalEditor
-  }
-
-  return rootEditor
 }
 
 /** Table insertion button with the legacy 5x5 hover picker used in the previous editor UI. */
@@ -155,51 +144,5 @@ export const MdxInsertTableButton = () => {
         </PopoverContent>
       </Portal>
     </Popover>
-  )
-}
-
-/** Text alignment controls restored to the markdown toolbar (left, center, right). */
-export const MdxTextAlignmentButtons = () => {
-  const isReadOnly = useCellValue(readOnly$)
-  const rootEditor = useCellValue(rootEditor$)
-  const editorInFocus = useCellValue(editorInFocus$)
-
-  const applyAlignment = (alignment: 'left' | 'center' | 'right') => {
-    const editor = getFocusedEditor(editorInFocus?.editorRef, rootEditor)
-
-    if (!editor || isReadOnly) {
-      return
-    }
-
-    editor.dispatchCommand(FORMAT_ELEMENT_COMMAND, alignment)
-  }
-
-  return (
-    <>
-      <ButtonWithTooltip
-        title={t('Align left')}
-        aria-label={t('Align left')}
-        disabled={isReadOnly}
-        onClick={() => applyAlignment('left')}
-      >
-        <PiTextAlignLeft size={16} />
-      </ButtonWithTooltip>
-      <ButtonWithTooltip
-        title={t('Align center')}
-        aria-label={t('Align center')}
-        disabled={isReadOnly}
-        onClick={() => applyAlignment('center')}
-      >
-        <PiTextAlignCenter size={16} />
-      </ButtonWithTooltip>
-      <ButtonWithTooltip
-        title={t('Align right')}
-        aria-label={t('Align right')}
-        disabled={isReadOnly}
-        onClick={() => applyAlignment('right')}
-      >
-        <PiTextAlignRight size={16} />
-      </ButtonWithTooltip>
-    </>
   )
 }
